[FIX] sap.ui.test.actions.EnterText - Event simulation succeeds if an element is not focusable
if a test is running in a background tab or the mouse is in the console of the developer tools, the EnterText action was not correctly triggering the focus events.
The FocusHandler of the core uses DOM events to attach to focusevents in EnterText i fired the events using jQuery so the events were not propagated to the Focushandler this results in uncalled sapfocusleave/enter functions.
With this change, the actual focus is not set on the control only the dom events are emulated
Worked fine for all modern browsers.
In IE11 the FocusEvent is not a constructor so i had to deal with the legacy syntax.
In PhantomJS there is no initFocusEvent so yet another special path
Tested it in Edge seems to be fine there

BCP: 1680124890
CR-Id: 012006153200001388182016
Change-Id: I92bb405987fbb4ef56c4ae270feaea682f3cd64f