[FEATURE] v2.ODataModel: getObject with select and expand parameters
In the ODataModel we store entity data in an internal structure, where
all expanded entities are replaced with references. To allow the
application to access the data on the client in a more convenient way,
this change introduces the possibility to define expand and select
options, to chose which properties/navigation properties to include in
the data.

BCP: 1680119838
Fixes https://github.com/SAP/openui5/issues/720

Change-Id: I3bc74c2b8795f4f5a9adaae8766c0af2eb01cab0
CR-Id: 012003146900000568972016