[FIX] ThemeCheck: custom.css check via pseudo element style
The current custom.css check was only working when the corresponding
rule (sapUiThemeDesignerCustomCss) was the first or second rule of
the library.css stylesheet. This should normally be the case as the
rule is also created as the first one via LESS mix-in
(see sap/ui/core/themes/base/shared.less).
In case some rules are prepended to the CSS, this check will fail.

As checking all rules (or a random number of rules) will be a
performance issue this change will use a different approach similar
to the inline theming parameters concept already in use.

Another advantage of the new approach is that it works cross-origin
(e.g. UI5 and/or theme is loaded from another server)

The old check has been kept in order to support old themes with a
newer runtime.

BCP: 0020751295 0000219741 2016
BCP: 0020751294 0000324874 2016

CR-Id: 012003146900000698142016
Change-Id: I07ba736c1d4fdb266d67863130f4297ebf00f117
(cherry picked from commit c610282000f3a0072d20f0c55c96d99b9fa4e119)