[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Allow preserve expanded state upon change of selected section
When the currently selected section becomes inavailable (due to being deleted or set to invisible), the ObjectPage automatically changes the selection to the first visible section and scrolls to that first visible section using "scrollToSection". The problem is that "scrollToSection" always brings sticky mode, even if we scroll to the first visible section. Added a check to avoid scrolling if the page is already expanded (i.e. first section is already visible)

Change-Id: Ifaa1441f87b56967e70132214c31668f69883708
BCP: 1680135596