[FIX] sap.m.NavContainer - Firefox state problem fixed
BCP: 1680140633
This change addresses a peculiarity of the Firefox browser with respect to page navigation calls that happen in a very short interval of one another.
It is possible that ".to()" is called during ".render()" (before the DOM has been flushed by the browser).
In this scenario, "this.getDomRef()" is still "false", therefore the ".to()" method returns early, expecting that ".render()" will render the last page in the stack.
However, ".render()" has already been called and will not be called again, and what is in the DOM is actually the old "current page".

Change-Id: Ife3b70054a5ddf54c5dfdfba2b79c60806f060d5
(cherry picked from commit c7f3030a8f26aeb6656ca0a4d83b8c24f1dff18f)
CR-Id: 012006153200001436422016