[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Allow preserve expanded state upon change of selected section
When the currently selected section becomes inavailable (due to being deleted or set to invisible), the ObjectPage automatically changes the selection to the first visible section and scrolls to that first visible section using "scrollToSection". The problem is that "scrollToSection" always brings sticky mode, even if we scroll to the first visible section. Added a check to avoid scrolling if the page is already expanded (i.e. first section is already visible)

Change-Id: I6ca1889e0b4ed810806de705dc430db1c1977f0e
CR-Id: 012006153200001416932016