[INTERNAL] Theming: remove workaround for IE9 style sheet limits
In case a library.css contained more than 4095 CSS rules, it has
been split into multiple files as IE9 and below had not been able to
parse them in a single style sheet.

IE9 also imposed a limit on the number of style sheets in a page.
jQuery.sap.includeStylesheet worked around that by adding imports to an
existing stylesheet instead of adding a new link tag after the limit was
reached.

As IE10 and below are no longer supported by UI5 1.40 and above, both
workarounds can be removed now.

Change-Id: I2baab4821f4570b10a5fca5e5e93fea31c682041