[INTERNAL] sap.ui.table.Table: New Keyboard behavior (Shift + Left/Right, Column Resizing) - Part 10
Implementation of the column resizing with Shift+Arrow Left to decrease a columns width, and Shift+Arrow Right to increase a columns width when a column header is focused, and this column is resizable.
In RTL mode pressing one of the key combinations on a column header has the opposite result.

Change-Id: Ia720e3fc15a1b19c464a7015f514ca84c6e39962