[INTERNAL][FIX] Added safe-check for displayable subsection in section
During layout calculations, and specifically in the _updateScreenHeightSectionBasesAndSpacer function, we can have a section *without any visible subsection*. This will be caused if we make a subsection invisible and relayouting occurs before invalidation. Added a safe-check for presence of subsection.
This is a downport change, as the same check is already added in more recent versions

Change-Id: I2efb912f5d79054a9dde08078ad6f055a61373fb
BCP: 1680231209
CR-Id: 012003146900000583672016