[FIX] sap.uxap.ObjectPageSectionBase - changes to subsections always update the anchor bar
When useIconTabBar=true, not all subsections are rendered initially.
As a result, the check "this.$().length" prevents the call to "_notifyObjectPageLayout()"
so changes to the subsections (for example title text) are not reflected in the AnchorBar.
This change replases the check if the subsection is rendered with whether
the ObjectPageLayout is rendered.

CR-Id: 012003146900000633022016
Change-Id: I32fe78f822179797955d26828cdda37d81b67fe4
BCP: 1670381177
Fixes: https://github.com/SAP/openui5/issues/1102
(cherry picked from commit e78c82553dcf095558a7cf48866e79c90d1c0f22)