[FIX] qUnit: Remove or refactor theme dependent unit tests
- Remove a test from Panel.qunit.html as it's depends on CSS background 
colors. These colors should be tested with visual tests, 
but not with qUnit

- Popover.qunit.html: Failing tests were checking the position of the 
popover. Popover's arrow is also taken into account for these tests and 
its size depends on the applied theme (arrow's size is hardcoded 
in qUnits). This is not testable. However we gain value of these tests 
if popover's positioning gets broken. We've adjusted arrow's hardcoded 
value in a manner it would throw an error if the popover's positioning 
gets broken, but it won't impact Belize or BlueCrystal.

BCP: 1670364572
BCP: 1670364440

Change-Id: Ifd3be270d8bea784df041dfe4aceba925fa47c5b