[INTERNAL] sap.m.ComboBox: update the selection before the change event is fired
Some application developers have the assumption that when the "change" event
is fired, not only the "value" should be updated,
(but also the value of the "selectedKey" and the "selectedItem" association).
This was the case for the desktop version, but not for the mobile version.

Now for consistency, the mobile version has been adapted.

Change-Id: I0d8fe0bc745dd9387135ad66502159c298bdfde1