[FIX] jQuery.sap.global: Added fix for Firefox sync/async XHR issue
As described in https://bugzilla.mozilla.org/show_bug.cgi?id=697151 ,
Firefox does wrongly execute event handlers of asynchronous requests,
while waiting on the response of synchronous requests. The workaround
for this issue is to create a proxy XHR, which uses a setTimeout to
delay the call of asynchronous event handlers, until the synchronous
XHR is completed.

BCP: 1670467873
Change-Id: I4c273436bd099fee1a523b3cd77e5f9a351eb81b
CR-Id: 012003146900000733212016