[INTERNAL] library-parameters.json: Option to track / disable sync XHR
6c62eb73906f4ad1f33ed0fa1a0989843c42054b introduced a new internal
config to track / disable sync XHR. With this change the Parameters
API will also respect the config and prevent / log sync XHR to
"library-parameters.json" files.

Furthermore a warning log has been added in case the API is used
before the core has been initialized. This could lead to the fact
a sync XHR is sent as the inline parameters are not available, yet.
Therefore the API should only be used e.g. in onBeforeRendering and
not generally in control classes or on control init.

Change-Id: Iafea844ffe04f84c1933d09374f4ed8c23fcde30