[INTERNAL] TabContainer: make _toTabStripItem private
The sap.m.TabStripItem class is marked as private. So it must not occur
in signatures of public or protected methods or the API doc will not be
complete. 

It seems that the _toTabStripItem method is used only internally but not
intended for other controls or for the application, so there seems to be
no reason to make it protected.

Change-Id: I39646bce0dd357698fde348fd94b785d782f1421