[FIX] sap.uxap.ObjectPageHeader - IE focus problem fixed
This patch addresses a loss of focus issue that happens on IE when an action button is made invisible.
Unlike other browsers, IE fires a focusout event when an element's visibility is changed to hidden,
even if it is changed back to visible in the same tick.
The code that changes action button visibility was introduced to address windows phone issues,
and is not needed on desktop, where focus matters.

Change-Id: I5950adb84c8e7b1e54a620c871e1ebfbdbdcf870