[INTERNAL] CompositeBinding: Work with array properties
Our controls do support array types for properties (e.g. string[]), but
these can only be bound, if the model data does contain an array. This
change enables the usage of a CompositeBinding for binding of array
properties, where yuou can define the array entries as parts of the
CompositeBinding. 

Change-Id: I7786d2da394988d50317c40627af9d7a8eaa3960