[FIX] sap.ui.fl UnstashControl - fix visible + optional move
The stashed marker control has no visible property, so running the change on a stashed control will throw an exception. The visible property if available is already modified in the setStashed method of the JSControlTreeModifier.

To harmonize the behavior of unhide and unstash, the option to move a stashed element directly in the unstash change is made optional.

Change-Id: Idff554147dcc25776e466434dc0c86df29116c5e