[INTERNAL][FIX] sap.ui.layout.ResponsiveFlowLayout - render manager is lazy obtained.
ResponsiveFlowLayout creates a (private) RenderManager in its constructor,
that is too early (in its contsructor).  As long as we support execution of
application code in the page head immediately after bootstrapping UI5
(synchronously), it might happen that control constructors are called before
the page DOM is ready. The FocusHandler is created only after that point in
time and only then it can be propagated to newly created RenderManager
instances.

BCP: 1670391482

Change-Id: Ida765538feee754164420adb1fb4b35304e81c2a
CR-Id: 012003146900000704172016