[INTERNAL] Component: Support OData Metadata/Annotations Cache Token
Cache tokens provided via the "asyncHints" API of the sap.ui.component
factory will be processed to enhance the model configuration of
"v2.ODataModel"s to make use of $metadata / annotation document caching.
In this case the "sap-language" needs to be provided as well because
switching the language should not reuse previously cached files which
could happen if the URI is the same.

Change-Id: Ic763ee606f3f506a60c28d4991a08e572b2c1113