[FIX] sap.ui.table.Table: fixes endless loop in IE if the table has no width
- If the visibleRowCountMode is "Auto", and the table width is smaller than or equal
the left + right border width, there is an endless-loop in IE. This happens since IE returns 0 for
the clientHeight of the table content when the clientWidth is 0.
This situation can happen when the table is set to invisible and its DOM shall be preserved.
If the table get a min-width, IE will return the correct clientHeight

Change-Id: I61c14f19daa2f6aa2c4539fda49b7113bc335777
BCP: 1670359747