[FEATURE] Set LTR content direction on initial rendering
The content direction was only set for RTL on initial rendering. If the
setting was changed later on, the html attribute was also set for LTR.
So, the behavior was inconsistent.

Since LTR is the default, rendering the attribute wouldn't be necessary,
but it was deemed more backwards-compatible, in case someone used a CSS
selector like html[dir="ltr"].

Change-Id: I430c0205c102b37f329e7523fc824c195d6e01d0