[FIX] sap.m.FacetFilter: when the last selected item is deselected list does not remain active; 'All' logic in the filter name is partially restored.
Filters that are initially displayed should stay on the UI, even if they are empty. Since the app designer has put them there by design, they are probably the most used filters.
 If a filter is set with a value and the user deselects the value, the filter should still be displayed.

Only if the user adds a new filter and does not add a value, it is supposed not to be shown.

In addition, if the user selects the all checkbox in the value selection, the filter should appear as "Filter name (ALL)" instead of "Filter name (234)".
That way, the user can see on the first glance if all values are selected, or not. The way it was implemented the user doesn't know if 234 are all filters or just a subset.

BCP: 1670412687
Change-Id: I41fa28b3b0a565559552b8c645f6dbb74f9cd7c4