[FIX] Configuration: improve locale and SAP language code handling
The session language has to be a BCP47 language to allow its usage in
HTTP headers, to be able to map it to CLDR locales and to SAP
translation keys (which basically are Java locale strings). 
This is reflected in the existing API of sap/ui/core/Locale,
sap/ui/core/Configuration etc.

However, the language information often is provided as sap-language
parameter, which should accept all SAP language codes as described in
http://scn.sap.com/docs/DOC-14377 . Not all these language codes are
valid BCP47 language tags, some not even syntactically (those
containing a digit). 

With this change, handling of the sap-language parameter became more
tolerant. Only when a parameter value can be interpreted as a BCP47
language tag (or when UI5 knows how to convert it - e.g. for 1Q, 2Q, ZH,
ZF), it will be used for the session language. Otherwise, it will be
remembered only as a 'SAP Logon language' setting and will not cause a
runtime error. Other language related parameters like sap-ui-language or
sap-locale no longer override the SAP Logon language. 

The runtime API 'setLanguage' still validates a new language against
BCP47 and fails in case of an error, but it now accepts a second
parameter sapLogonLanguage, which is only validated against the SAP
Language code definition (two letters or digits) and also stored as SAP
Logon language. It is up to the caller of that API to keep the two
language settings consistent.

The SAP Logon language setting can be retrieved with the method
Configuration.getSAPLogonLanguage and can be used in URLs that target
SAP systems. The method in sap/ui/core/Locale with the same name has
been deprecated as it doesn't know about the new configuration setting.

(cherry picked from commit fdb494f333098627371fece1ae51348e0ae50275)
Change-Id: Id5d80de16701202c5c171f5af8e41c0b7256b942
BCP: 002075129500003012162016
CR-Id: 012006153200001437312016