[FIX] View: Fix cloning of template listeners
When cloning a ManagedObject we clone its event listeners as well.
If an instance attaches to events during initialization, those event
listeners will get overwritten in case the instance is created in a clone procedure.
Such instances end up with the event listeners of the template instance
instead of its own.

In case of views this leads to view clones firing events on the
template controller instance instead of their own controller. E.g.
onAfterRendering is always called in the context of the template
instance.

This fix will ensure that event listeners on a cloned view which should
be called on the clones controller, will be called on that.

Change-Id: Ifcf879a430fef63e51d349b83c8b9dd50f65f751
BCP: 1670506962