[FEATURE] Component: append SAP params to OData and Manifest requests
#1: Detection of SAP parameters (sap-client, sap-server, sap-system) in
Configuration from URI parameter or META tag of the document. If the
client/server/system cannot be determined from the URL it can be read
from a META tag in the HTML page:

<META name="sap-client" content="100">
<META name="sap-server" content="abc">
<META name="sap-system" content="sys">

#2: Enhanced the Manifest loading to add the sap-client URL parameter to
the request if present as SAP parameter in the document.

#3: Improved the backend context token logic to append the context token
only in the case a sap-language is present and the sap-client parameter
is available and is matching the sap-client parameter of the OData
entity (either datasource or annotation uri). Only in this case the
context token can be added to the backend requests.

Change-Id: Iaddf1e793536a69942772b39f3cd440fe7ea31e2