[INTERNAL][FIX] sap.m.PlanningCalendar: Navigation with left arrow in week view from non-summer to summer week works now.
When current week is the first non-summer (e.g. 2 Nov - 8 Nov 2014, US calendar) and the user focuses the 2nd of Nov and clicks arrow left, then
the previous week was wrongly positioned (e.g. 25 Oct-31 Oct) and the focus is lost.

Expected is that the new week should be exactly 7 days before the current (e.g. 26 Oct - 01 Nov), and the focus should be at the last day
of the new week (1st of Nov).

Change-Id: Icb092f7300d4110a964ac734467860405e49bbbd