[FIX] Firefox/Safari: Synchronous XHR and Promises
When a synchronous XHR is triggered within a Promise callback, Firefox
does not prevent other Promise callbacks to be executed while the
request is ongoing. This change does catch all Promise callbacks called
while a synchronous request is in progress and delays their execution on
after the synchronous request is completed.

Safari has a broken Promise implementation, where it is possible that
timeouts are executed before promise callbacks or even in between a
promise chain. Falling back to the Promise polyfill for Safari, as we do
for Edge.

Change-Id: I8f005d9e649908a020a4005bb44dcce7e140cbdd