[FEATURE] core/Component: Introduce settings of services into manifest
Currently the concept of UI5 services is lacking a proper way to include
settings.

With this change a new schema of the app descriptor is introduced, in order
to include the corresponding service settings. The idea is to group all
service settings into a section of its own. This group of settings would
be pased to the service for instantiating.

Change-Id: Ia31aff71f88bc29e30c569bf14f4aa450bb48c12