[INTERNAL] Core: improve waitForTheme on Safari
The conditions implemented by the ThemeCheck to detect whether CSS has
been loaded and applied don't always match the visual state. Esp. with
empty caches, a flash of unstyled content (FOUC) occasionally occurred.
Now adding a small delay as a last resort until a better condition has
been found.

Additionally, the UI5 logo in the SDK is rendered only when the UI5
rendering starts and not earlier on DOM ready.

Last but not least, fix a typo in one of the lists of supported themes.

Change-Id: I7cb0c626ecb9a30db5518b44ab598e7968caa11f