[FIX] sap.ui.core.util.ExportTypeCSV: Don't escape number only formulas
Numbers like -123 have been escaped as a hyphen ("-") as first char may
indicate a formula which should be escaped due to security reasons.
As this causes issues for negative numbers (and potentially more cases)
the regular expression has been extended to check if the whole string is
basically just a number which does not need to be escaped.

Furthermore the pattern to detect content which needs to be enclosed
with double quotes has been extended to prevent issues with common
separator chars which might be automatically used by programs.

CR-Id: 012006153200001516782016
Change-Id: Iea9b4cf61447a077943cabb18a202d034f4fd5c1
BCP: 1670454040
(cherry picked from commit 3af166ff78b2697e9ffed0dca7584d7226eb26b3)