[INTERNAL] sap.f.FlexibleColumnLayout - API simplified, semantics introduced
This change simplifies the API of the control by merging several properties (showMidColumn, showEndColumn, fullScreenColumn) and the public method "setLayout" into just one property (layout).
A semantic wrapper on top of FlexibleColumnLayout is also introduced. It takes advantage of the "layout" property and hints the user about the expected UI behavior of action buttons and links.

Change-Id: Ieb4eead9a17671e08435047f6c29625d15e26347