[FEATURE] changeHander registration can be done in one file per control
Within sap.ui.fl changes on controls can be registered via the metadata
of a library. This is now not only possible by providing a mapping of
actions (changeTypes) and Classes (changeHandlers) within the library.js
itself but also as a reference in form of a module path. This module has
to have a .flexibility.js file type.
I.e. the registration of "sap.ui.layout.form.SimpleForm" is done via the
string "sap/ui/layout/flexibility/SimpleForm" which results in the
loading of "sap/ui/layout/flexibility/SimpleForm.flexibility.js".

Change-Id: If1a4376c39267312b1d9d7d5b53a95058002c44d