[FEATURE] Component: explicit usage declaration for reuse components
Reuse components being used inside a component need to be explicitly
declared in the manifest in order to create them with a dedicated
factory function inside the component. This ensures that the nested
components will get the proper scope of the owner component.

This feature also allows flexibility to exchange nested components by
just modifying the descriptor/manifest. In addition, this feature is
also performance relevant as the usage of nested components can be
declared and loaded later via the createComponent API.

Change-Id: Ie4882ecd919533285e56acad0baa9aef67372072