[FEATURE] sap.ui.core.util.ExportType: Add 'byteOrderMark' property
Adding a property to decide on whether a byte order mark should be added
when downloading an expoerted file via the
"sap.ui.core.util.Export#saveFile" method.

Also added the corresponding parameter to the
"sap.ui.core.util.File.save" function.

The previous behavior has not been changed for compatiblity reasons.
If the property/parameter is not set/given the byte order mark will be
added for "csv" files by default. Otherwise it will not be added.

BCP: 1680269676
Change-Id: I5798ade25f67c973b6c2f6994646d69c21ee128a