[FIX] ThemeCheck: Fix issues when switching theme / improve logging
ThemeCheck.checkStyle had issues when the theme has been switched during
runtime (via sap.ui.getCore().applyTheme). The check could not identify
a stylesheet which is either empty or not found (404) as the custom
"data-sap-ui-ready" attribute was not added to the <link> element.
This issue is because switching the theme did only update the element's
"href" property which causes re-loading the new theme but not re-firing
the "load/error" events. This is fixed by calling
"jQuery.sap.includeStyleSheet" which then replaces the existing element
with a new one.

The "load/error" events of <link> tags seem to be supported by all
relevant browsers:
IE 11, Edge, Chrome, Firefox, Safari 10, Safari Mobile (iOS 8.4, 9.2,
10.0)

BCP: 1680246136
Change-Id: I9496a86b55985dd129073fe023ab3868c72bad83