[FIX] mvc/Controller: support controller extension in factory function
With the enablement of async controller extensions the extendIfRequired
call has been moved into the View when connecting the View and the
Controller. This lead to a regression which prevents the controller
extensibility when using the Controller factory function only to create
a new Controller instance. This fix restores the old behavior.

Change-Id: Ia9dfb615d68693b2ef00d7717da80430c3c51082
BCP: 002075129500004107382016
CR-Id: 012006153200001502952016