[FEATURE][INTERNAL] fl/Utils determines if personalization is in place
The function sap/ui/fl.isApplicationOfControlPersonalized was introduced
checking if any personalization for a given control/component is in
place.
This is required to determine if RTA can be initialized for a user
making adaptations in a layer above the user layer. If a personalization
is in place the personalization has to be removed.

Change-Id: I007f72084bc83fb102b9eadc24782e92d835d9c6