[FIX] sap.ui.table.Table: Ignore the additional empty row at the end in the vertical scrolling feature.
When variable row heights are enabled an additional row is rendered for pixel based scrolling.
When the row count was equal to the visible row count this additional empty row caused the vertical scollbar to be visible,
even though not required. When calculating the row heights delta this row will be ignored.

Change-Id: Ib98999c2e22bed218da85862588bd80b442c94de
BCP: 1680285544
CR-Id: 012006153200001513462016