[FIX] ODataModel: Normalize entity URL encoding and number values
In v2.ODataModel entries are accessed on the local cache based on the
canonical URL. There are two cases, where this can go wrong: The
encoding of string properties might be different and for
Decimal/Int64/Float/Double the appended character can be upper or lower
case.

This change makes sure that created entities do always use lower case
characters for number types and adds a normalize method to the
v2.ODataModel to adapt all canonical URLs received from the server to
match the expected format.

Change-Id: Id0864afbec68a00af48d8a2baaf4be22a8efca8c