[FIX] ManagedObject: fail early when an aggregation cycle is created
Before this change, isInvalidateSuppressed failed with a stack overflow
which is hard to analyze. Running up the hierarchy just for the check
might be expensive for bigger updates of larger object trees. But the
implementation of isInvalidateSuppressed has a similar complexity, is
already called in setParent and is not known to cause performance
issues.

Related to https://github.com/SAP/openui5/issues/1264 .

Change-Id: I60817928fd31c96b02d15c708dace00b0bf80456