[INTERNAL][FEATURE] Flexibility changes can be discarded by control id
JIRA: CPOUIFDALLAS-1185

The deletion of all flexibility changes for a specific control is
required to remove the SmartLink personalization as well as the current
development cycle. Therefore a function was added to retrieve these
changes and delete them.
A flag is used to determine if the personalization should be removed or
the actual development layer.

Change-Id: I6bb576868b87a1ce52adf44bbfc2f904446b83ed