[INTERNAL] grunt: Serve sap-ui-version.json based on configured libs
As the sap-ui-version.json file within the testsuite project is anyway
only relevant for local development it makes more sense to dynamically
serve it based on the configured libraries (which will then also include
libraries added via "--config-extension" scripts).

This also reduces the configuration effort when adding a new library.

Change-Id: I3b6d529316f5d9946acaf67002498594dc1d105e