[INTERNAL] sap.m.semantic.SemanticPage: Added action elements to library.js
Some of the semantic action elements (added in a later stage) we not respectively added to library.js.
As a result of that, no lazy stub was created for them and also had to be explicitly required.
Added them to the respective elements list in library.js of sap.m.

Change-Id: Ic12b33cb83bc3d9188ae64136ec75028d7641ec0