[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Sections dropdown items will have their sectionId as key
The ObjectPage has a sections dropdown [that appears in phone view in anchorBar overflow] that allows to select any of the available sections.
The keys of these dropdown items were the ids of the associated anchorBar buttons (assuming there is always an anchorBar button associated with a section).
However, when "showAnchorBarPopover" is "false", then there are NO anchorBar buttons associated with the SUBsections.
Fixed by making the keys of the (sub)section-items to be the ids of the actual associated (sub)sections. No intermediary object (as another section button) is needed.

Change-Id: Ie3d36ff8a0f4a4452e6b08652ef9b333924ac5b2
BCP: 1680335022
(cherry picked from commit bf88294d24cbf5d9bde943b3e29987ca5faf5d0c)
CR-Id: 012006153200000017852017