[INTERNAL] sap.f.FlexibleColumnLayout - nav containers creation delayed
This change optimizes the initial rendering of a non-empty FlexibleColumnLayout.
The three nav containers are not rendered initially, but are flushed during onAfterRendering.
This ensures that controls will not have to do double work due to the resizing of the columns,
which in turn can only take place after the control width is known.

Change-Id: I3c627ec23216098a02da410f2e0840b72a8e14f4