[FIX] sap.m.ViewSettingsDialog - calling method removeAllSortItems() does not throw an exception if no items exist.
When sortItems aggregation is empty, removeAllSortItems wrongly assumed that
there is always an internal list for sorting available. This list is usually created when
at least one sort item exists.

There is an optimization logic that prevents internal (sorting) list recreation and allow for reuse of it.
It creates a map that usually should keep a reference to the corresponding internal (sorting) list.
The removal of (sort) items logic, wrongly assumed that this list should be always available.
If no (sort) items exists, it will be never available.

Change-Id: I3425328795ed3548c6570a0661b14275521c12b8
BCP: 002075129400000165292017