[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Restore selected state upon section deletion
When the selected section was deleted, the selection was loast in some cases (when in iconTabBar mode and scroll position is 0).
Made sure selection is set to the first visible section if the previously selected section is inavailable. Enabled the respective qunit.

Change-Id: I7cbf0ff56770cd05714cd0847f0ff1aa5f9a7b1e
BCP: 1680286545
(cherry picked from commit 87dd7d8dbd940d32049291deb123d1f22625ba25)
CR-Id: 012006153200000031642017