[INTERNAL] grunt: Serve sap-ui-version.json based on configured libs
As the sap-ui-version.json file within the testsuite project is anyway
only relevant for local development it makes more sense to dynamically
serve it based on the configured libraries (which will then also include
libraries added via "--config-extension" scripts).

This also reduces the configuration effort when adding a new library.

Change-Id: I6f65cf501eedbcc5d2f717138866b6c4955565a9
CR-Id: 012006153200000014452017