[INTERNAL] grunt: Serve sap-ui-version.json based on configured libs
As the sap-ui-version.json file within the testsuite project is anyway
only relevant for local development it makes more sense to dynamically
serve it based on the configured libraries (which will then also include
libraries added via "--config-extension" scripts).

This also reduces the configuration effort when adding a new library.

Change-Id: I425f09f78e9b9ccc3acddad4d7dd1971331aeabc
CR-Id: 012006153200000014442017