[FEATURE] Configuration: Technical switch for multiple animation modes
This change introduces a new string property "animationMode" for
different animation modes. The animation modes are validated against
an enumeration list.
The current valid values are "full", "basic", "minimal" and "none".
The old animation property is deprecated. The value "none" turns the old
animation property to "off". The other values turn the animation to
"on". This ensures backward compatibility. If an animation mode is
provided it overwrites an existing animation setting.
The method "setAnimationMode" of the configuration changes the mode
and updates the new HTML attribute "data-sap-ui-animation-mode=<mode>"
to the corresponding value. The configuration class is now able to
validate against different enumeration objects (e.g. AnimationMode).

Change-Id: I31dfc7de66650b07f1e5872ba7e25ec8f302892c