[INTERNAL] sap.m.ObjectMarker - container-aware media API used
This change makes ObjectMarker render correctly in containers that propagate "contextualWidth".
A prerequisite for this is that _adjustControl is called when the control is part of the control tree (onBeforeRendering),
so that contextual width is available (if inside a media-aware container).
This in turn removes the need for custom setters.

Change-Id: I909fb44ccdd82c6235209b58cf43e516e865cd29