[FIX] ManagedObject: _removeChild should not invalidate on its own
If the given child object is found in the aggregation, _removeChild
dispatches to the matching mutator method to remove it. Calling the
mutator will already take care of invalidation. The mutators might also
implement there own logic to activate / suppress invalidation. Calling
invalidate() again inside _removeChild would not 
only be a redundant call but also would ignore that logic.

One important use case where this unnecessary invalidation occurred
before was the move of an object within the "dependents" aggregation.

Change-Id: I99941010da22841a65a864c9cd06156629ace555