[FIX] sap.ui.core.IconPool: Use sap.ui.getCore() instead of AMD export
When the IconPool is required during core bootstrap there was an
exception when accessing the local "core" variable which is undefined as
the actual dependency (sap/ui/core/Core) is not fully exported at this
point in time.
Therefore "sap.ui.getCore()" should be used instead which is defined
already.

Change-Id: I42cfcaa6caaf815dcffcc647b8f90bcc78fe0d01
BCP: 1780046575
BCP: 1780055543
BCP: 1780052096
BCP: 1780055143
BCP: 1780055706