[FIX] ThemeCheck: Fix Cross-Origin issues in Firefox
With 5fa1c3250da6ffe998f39fa3c6f0111135d9d722 the ThemeCheck has been
improved in general but has also caused an issue in the custom css
check which did not happen before.
Firefox is throwing an exception when accessing the "sheet.cssRules"
object of a stylesheet which is loaded via CORS.
This is caught in the general stylesheet check but not when checking
the rules for the custom css marker.

CR-Id: 012006153200000044652017
Change-Id: If259a0eb669854995e1ed5571964f54ae5bdea9b
BCP: 1780051923