[FIX] ThemeCheck: Fix Cross-Origin issues in Firefox
With 73757b1d7c3298ff65894a9fff08ab2c7328e524 the ThemeCheck has been
improved in general but has also caused an issue in the custom css
check which did not happen before.
Firefox is throwing an exception when accessing the "sheet.cssRules"
object of a stylesheet which is loaded via CORS.
This is caught in the general stylesheet check but not when checking
the rules for the custom css marker.

Change-Id: If259a0eb669854995e1ed5571964f54ae5bdea9b
BCP: 1780051923