[FIX] ThemeCheck: Fix Cross-Origin issues in Firefox
With d5952034728f8e03518affaf74b233edd9198d2d the ThemeCheck has been
improved in general but has also caused an issue in the custom css
check which did not happen before.
Firefox is throwing an exception when accessing the "sheet.cssRules"
object of a stylesheet which is loaded via CORS.
This is caught in the general stylesheet check but not when checking
the rules for the custom css marker.

CR-Id: 012006153200000044642017
Change-Id: If259a0eb669854995e1ed5571964f54ae5bdea9b
BCP: 1780051923