[FIX] ThemeCheck: Fix Cross-Origin issues in Firefox
With d904b3c4bde88b28e170d2d1ea06ad764f9e0ff3 the ThemeCheck has been
improved in general but has also caused an issue in the custom css
check which did not happen before.
Firefox is throwing an exception when accessing the "sheet.cssRules"
object of a stylesheet which is loaded via CORS.
This is caught in the general stylesheet check but not when checking
the rules for the custom css marker.

CR-Id: 012006153200000044632017
Change-Id: If259a0eb669854995e1ed5571964f54ae5bdea9b
BCP: 1780051923