[FIX] v2.ODataListBinding: Skip data evens for aborted requests
If refresh, sort or filter is called on the binding, while a request is
still in progress, the current request is aborted and a new request is
sent to the server. In the past, for the aborted reqest the
"DataReceived" event was fired and the new request fired a new
"DataRequested" event. This made it hard for the application to
distinguish between real received data and aborted request with no data.
Now DataReceived event is omitted for the aborted request, as well as
DataRequested for the following new request, so for the application the
abortion is transparent, it does just see a single DataRequested event
followed by a single DataReceived event in the end, no matter how often
the request was aborted and restarted in between.

BCP: 1680064949
Change-Id: I9819fb43ad1955fcd98241ba85b3e8d08db85a4d
CR-Id: 012006153200000063942017