[INTERNAL] RewriteFilter: header handling / no rewriting when encoded
The header handling has been changed to be case insensitive like the
official specs. As this was a copy a previous experiment I just ignored
this in the first iteration.

Once a response has a content-encoding: gzip or any other encoding value
set the RewriteFilter will just ignore that responses. This is the case
when handling remote resources of the proxy like OData calls or when
loading UI5 resources from CDN. But they must not be handled by the
RewriteFilter as it has been implemented for the local development
scenario (in combination with the ResourceServlet) only.

Change-Id: I57c3796a63e17da305f123cd6d0d8926cf92953e