[FEATURE] ComponentContainer: support prefixing of component id
Introduces a property (autoPrefixId) on the ComponentContainer to allow
prefixing the Component ID when specified in the settings property. This
feature is useful when using the ComponentContainer in a declarative
view (like the XMLView) and defining a Component with an ID for the
ComponentContainer within the settings property. The ID of the Component
will then be prefixed with the ID of the ComponentContainer.

Change-Id: Iedef7c46a9907481a9f7edf01af73a7827e18760