[FIX] sap.m.PlanningCalendar-tap on appointment on mobile should not fire rowSelectionChange.
When the user clicks on appointment in a row that is not selected, for
mobile devices this used to throw rowSelectionChange event.

The fix details:
-CalendarRow click handler is removed, the same logic is put inside
'tap' event handler

-Tap event was not called for desktop. As tap event is a common event
for both desktop(click) and mobile (touch) devices the
mousedown/mouseup, touchstart/touchend should happen for one and the
same target dom element, otherwise the event won't be fired.
This wasn't the case for CalendarRow, because styling the focus
inserted additional dom element (pseudo :after) which actually
is being tha target element at mouseup. So the user
'mousedown'-s on the appointment dom element, but 'mouseup'-s
at the:after element. To fix this, 'pointer-events: none' added for
the additional dom element (:after).

BCP: 002075129500004765332016

Change-Id: I16ff2041ec344aefea98059e17c3d6299873833c