[INTERNAL] sap.m.Page - belize Standard background changed
By default backgroundDesign=Standard is not implemented by themes, thus mapping to Transparent.
This commit makes backgroundDesign=Standard map to Solid for the belize theme when in FLP only.
The change is needed so that pages without an explicitly set backgroundDesign property
can provide a solid background, in case no other control underneath provided it.

This behavior was already in place for the sap_belize_plus theme. The implementation is now moved
to sap_belize, so that it is applied to both sap_belize and sap_belize_plus.

Change-Id: I04b9aa15651caada68d23d88bb4499d09bd4cd63
(cherry picked from commit d63c3b1f0453a7ac6dfd271c2ede228ac796732b)
CR-Id: 012006153200000083172017