[INTERNAL][FIX]sap.uxap.ObjectPageHeader: Removed duplicate image
Ensured that a single image control (or a single image placeholder control) is rendered only once in the view.

The existing ObjectPage API requires:
- one image (or image placeholder) in the headerTitle and
- another image (or image placeholder) in the headerContent
(because the API allows both the image in headerTitle and the image in headerContent to be displayed at the same time).
Therefore fixed by ensuring that both headerTitle and headerContent create their own image instances.
Added qunits.

Change-Id: I7b9acee16565f62ae802a206e878b25e8540ae5b
BCP: 1780046278