[FIX] sap.ui.dt register overlays at plugins only when dt is synced
We have to ensure that overlays have designtime metadata and control set,
as well as having overlays for child and parent controls. Therefore we are no longer registering the
overlays directly on the plugins when it is created, but postphone it until all overlays that should be
visible are created completely (designtime synced).

Change-Id: I7ba2cbff349a5db8517e55d09c2bba6dac2cd490
BCP: 1780129678
CR-Id: 002075125900000217092017
(cherry picked from commit e2a9bb33b85ff068596aca282676135a6d41b8b3)