[FEATURE] sap.ui.core.util.Mockserver: Support URL Param "search"
SAP uses in addition to the OData v2 standard parameters two additional
URL parameters: "search" and "search-focus". "search" is used to provide
a search string to the backend, "search-focus" is used to provide the
primary property where the data should be searched. 
Although  the search logic implementation is backend-specific the
mockserver supports now a "slim" search ("startsWith" on the given
property) to fulfill demo purposes. 

Change-Id: I115d86da3ba4cfbb980b0a1272d511d5d9fc2542