[INTERNAL][FIX] m.PlanningCalendar: Today button state test corrected
The week view has the side effect of changing its startDate.
So after TODAY is given as a startDate, the modified startDate may be
in the previous month. Thus the unit test incorrectly checks if the 
button state is correct for the previous month.
One unfortunate example is TODAY===01/03/2017, where the modified
startDate is 27/02/2017.
The solution is to not use the week view for this test.

Change-Id: I3e276b8796c94740323e5589948c4ac77e2eb209
BCP: 1770034872
(cherry picked from commit f7cc7e4867e89896b59c41d49eee4eb03420af78)
CR-Id: 002075125900000357992017